Updated MDaaS to pass testing steps to Run_Multi-Part_Simulation.bash

+++ gemsModules/complex/glycomimetics/services/Build_Selected_Positions/logic.py
@@ -28,7 +28,11 @@ def execute(inputs: Build_Inputs) -> Build_Outputs:
     else:
         log.debug("Copying scripts")
         # Note, we do not execute the copies, they are only for reference at this moment.
-        os.system(f"cp -r {GlycoWebtool_path}/scripts {project_dir}")
+        # check glyco exists
+        if not GlycoWebtool_path.exists():
+            log.warning(f"Warning: GlycoWebtool path {GlycoWebtool_path} does not exist. Cannot run Glycomimetics.")
+        else:
+            os.system(f"cp -r {GlycoWebtool_path}/scripts {project_dir}")
     
         # Create the input.txt for the Glycomimetics service.
         # Note: We only allow one Selected Position at a time in the API for now.
